card_gen/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
openai==1.71.0
python-dotenv==1.0.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==1.0.1 